decouple storage from computation and make sure that floats are not used for computation in core code
git-svn-id: https://svn.salilab.org/imp/trunk@2105 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ kernel/include/base_types.h
@@ -21,7 +21,7 @@ IMP_BEGIN_NAMESPACE
  */
 
 //! Basic floating-point value (could be float, double...)
-typedef float Float;
+typedef double Float;
 
 //! A pair of floats
 typedef std::pair<Float, Float> FloatPair;
